implementing MaxScore

+++ src/main/java/it/unipi/mircv/File/LexiconHandler.java
@@ -77,9 +77,9 @@ public class LexiconHandler{
     public int getCf(ByteBuffer dataBuffer) {
         return dataBuffer.position(TERM_BYTES_LENGTH + OFFSET_BYTES_LENGTH + DOCUMFREQ_BYTES_LENGTH).getInt();
     }
-    public int getTermUpperBoundScore(ByteBuffer dataBuffer) {
+    public float getTermUpperBoundScore(ByteBuffer dataBuffer) {
         return dataBuffer.position(TERM_BYTES_LENGTH + OFFSET_BYTES_LENGTH
-                + DOCUMFREQ_BYTES_LENGTH + COLLECTIONFREQ_BYTES_LENGTH).getInt();
+                + DOCUMFREQ_BYTES_LENGTH + COLLECTIONFREQ_BYTES_LENGTH).getFloat();
     }
     public int getOffsetSkipDesc(ByteBuffer dataBuffer){
         return dataBuffer.position(TERM_BYTES_LENGTH + OFFSET_BYTES_LENGTH
